Tables: Move column definition to separate file
