+++ src/widgets/reports/PotentialLocales.tsx
@@ -7,7 +7,8 @@ import { LocaleSeparator, ObjectType } from '@features/page-params/PageParamType
 import usePageParams from '@features/page-params/usePageParams';
 import { getSortFunction } from '@features/sorting/sort';
 import { SortBy } from '@features/sorting/SortTypes';
-import ObjectTable, { ValueType } from '@features/table/ObjectTable';
+import ObjectTable from '@features/table/ObjectTable';
+import TableValueType from '@features/table/TableValueType';
 
 import { CensusData } from '@entities/census/CensusTypes';
 import { LanguageCode, LanguageData } from '@entities/language/LanguageTypes';
@@ -152,7 +153,7 @@ const PotentialLocalesTable: React.FC<{
         {
           key: 'Population',
           render: (object) => object.populationSpeaking,
-          valueType: ValueType.Numeric,
+          valueType: TableValueType.Numeric,
           sortParam: SortBy.Population,
         },
         {
@@ -160,7 +161,7 @@ const PotentialLocalesTable: React.FC<{
           render: (object) =>
             object.populationSpeakingPercent &&
             numberToFixedUnlessSmall(object.populationSpeakingPercent),
-          valueType: ValueType.Numeric,
+          valueType: TableValueType.Numeric,
           sortParam: SortBy.PercentOfTerritoryPopulation,
         },
         {
@@ -170,7 +171,7 @@ const PotentialLocalesTable: React.FC<{
             numberToFixedUnlessSmall(
               (object.populationSpeaking * 100) / (object.language?.populationEstimate ?? 1),
             ),
-          valueType: ValueType.Numeric,
+          valueType: TableValueType.Numeric,
           sortParam: SortBy.PercentOfOverallLanguageSpeakers,
         },
         {
