+++ src/widgets/tables/columns/LanguagePopulationColumns.tsx
@@ -1,5 +1,6 @@
 import { SortBy } from '@features/sorting/SortTypes';
-import { TableColumn, ValueType } from '@features/table/ObjectTable';
+import TableColumn from '@features/table/TableColumn';
+import TableValueType from '@features/table/TableValueType';
 
 import { LanguagePopulationEstimate } from '@entities/language/LanguagePopulationEstimate';
 import LanguagePopulationFromDescendents from '@entities/language/LanguagePopulationFromDescendents';
@@ -17,7 +18,7 @@ export const LanguagePopulationColumns: TableColumn<LanguageData>[] = [
       </>
     ),
     render: (lang) => <LanguagePopulationEstimate lang={lang} />,
-    valueType: ValueType.Numeric,
+    valueType: TableValueType.Numeric,
     sortParam: SortBy.Population,
     columnGroup: 'Population',
   },
@@ -25,7 +26,7 @@ export const LanguagePopulationColumns: TableColumn<LanguageData>[] = [
     key: 'Population (Direct)',
     description: 'This comes from other language databases (citations still needed).',
     render: (lang) => lang.populationCited,
-    valueType: ValueType.Numeric,
+    valueType: TableValueType.Numeric,
     isInitiallyVisible: false,
     sortParam: SortBy.PopulationAttested,
     columnGroup: 'Population',
@@ -35,7 +36,7 @@ export const LanguagePopulationColumns: TableColumn<LanguageData>[] = [
     description:
       'Some of these languages may have data from constituent dialects/locales. They have been added up here.',
     render: (lang) => <LanguagePopulationFromDescendents lang={lang} />,
-    valueType: ValueType.Numeric,
+    valueType: TableValueType.Numeric,
     isInitiallyVisible: false,
     sortParam: SortBy.PopulationOfDescendents,
     columnGroup: 'Population',
@@ -49,7 +50,7 @@ export const LanguagePopulationColumns: TableColumn<LanguageData>[] = [
       </>
     ),
     render: (lang) => <LanguagePopulationFromLocales lang={lang} />,
-    valueType: ValueType.Numeric,
+    valueType: TableValueType.Numeric,
     isInitiallyVisible: false,
     columnGroup: 'Population',
   },
