+++ src/widgets/tables/columns/LanguageVitalityColumns.tsx
@@ -1,5 +1,6 @@
 import { SortBy } from '@features/sorting/SortTypes';
-import { TableColumn } from '@features/table/ObjectTable';
+import TableColumn from '@features/table/TableColumn';
+import TableValueType from '@features/table/TableValueType';
 
 import { LanguageData } from '@entities/language/LanguageTypes';
 import LanguageVitalityCell from '@entities/language/vitality/LanguageVitalityCell';
@@ -10,6 +11,7 @@ export const LanguageVitalityColumns: TableColumn<LanguageData>[] = [
     key: 'Vitality: Metascore',
     render: (lang) => <LanguageVitalityCell lang={lang} type={VitalitySource.Metascore} />,
     sortParam: SortBy.VitalityMetascore,
+    valueType: TableValueType.Enum,
     columnGroup: 'Vitality',
   },
   {
@@ -17,12 +19,14 @@ export const LanguageVitalityColumns: TableColumn<LanguageData>[] = [
     render: (lang) => <LanguageVitalityCell lang={lang} type={VitalitySource.ISO} />,
     sortParam: SortBy.VitalityISO,
     isInitiallyVisible: false,
+    valueType: TableValueType.Enum,
     columnGroup: 'Vitality',
   },
   {
     key: 'Vitality: Ethnologue 2013',
     render: (lang) => <LanguageVitalityCell lang={lang} type={VitalitySource.Eth2013} />,
     sortParam: SortBy.VitalityEthnologue2013,
+    valueType: TableValueType.Enum,
     isInitiallyVisible: false,
     columnGroup: 'Vitality',
   },
@@ -30,6 +34,7 @@ export const LanguageVitalityColumns: TableColumn<LanguageData>[] = [
     key: 'Vitality: Ethnologue 2025',
     render: (lang) => <LanguageVitalityCell lang={lang} type={VitalitySource.Eth2025} />,
     sortParam: SortBy.VitalityEthnologue2025,
+    valueType: TableValueType.Enum,
     isInitiallyVisible: false,
     columnGroup: 'Vitality',
   },
